Fixed warnings
Added preprocessor definition "_HAS_STD_BYTE=0"
Updated to c++17 and C17.
Built and compiled on x64 and x86, no warnings

+++ TP3/Infirmier.cpp
@@ -66,7 +66,7 @@ void Infirmier::ajouterHeuresTravaillees(unsigned heures){
 }   
   
 float  Infirmier::calculerSalaire() const {
-	float salaire = heuresTravaillees_ * tauxHoraire_ + listeChambre_.size() * tauxHoraire_ / 20.0;
+	float salaire = static_cast<float>(heuresTravaillees_ * tauxHoraire_ + listeChambre_.size() * tauxHoraire_ / 20.0);
 	return  salaire;
 }
